feat(logging): Implement SPIFFS-based logging system
- Added LogManager for managing log levels and queuing log messages.
- Introduced LogFile for handling log file creation, rotation, and reading.
- Created JsonBuilder and JsonParser for efficient JSON serialization and deserialization.
- Updated WebServerManager to include access control and logging functionalities.
- Removed old Logger class in favor of the new logging system.

+++ src/hal/MPU6050Driver.h
@@ -3,6 +3,7 @@
 
 #include <Arduino.h>
 #include <Wire.h>
+#include "../config/Config.h"
 
 class MPU6050Driver {
 public:
@@ -14,7 +15,7 @@ public:
     void readRaw(int16_t& ax, int16_t& ay, int16_t& az, int16_t& gx, int16_t& gy, int16_t& gz);
     
     void setAccelRange(uint8_t range);   // 2, 4, 8, 16
-    void setGyroRange(uint8_t range);     // 250, 500, 1000, 2000
+    void setGyroRange(uint16_t range);    // 250, 500, 1000, 2000
     void setDLPF(uint8_t bandwidth);      // 0-6 (260Hz to 5Hz)
     
     void enableMotionInterrupt(bool enable, float threshold = 0.5f);
